==> 2.4 Classification_SC3.Rmd <==

# 1 Preparation
## 1.1 Library
```{r}
library(Seurat)
library(SingleCellExperiment)
library(SC3)
library(ggplot2)
library(scater)
```

## 1.2 Bulk data: Count
```{r}
tpm_122_Lumbar_noN_CV <- read.xlsx("/linp/linp/IDD_classification/Formal Analysis/1 Primary matrix/TPM_122_Lumbar_noN_CV_5.xlsx")
```

```{r}
# count_120_Lumbar_noN_CV <- count_120_Lumbar_noN_CV[!count_120_Lumbar_noN_CV$Genes %in% c('HBG1','HBE1','HBG2','HBD','HBQ1','HBM','HBA1','HBA2','HBB','HBP1'),]

tpm_122_Lumbar_noN_CV <- tpm_122_Lumbar_noN_CV[order(tpm_122_Lumbar_noN_CV$CV,decreasing = T),][1:8000,]

Genes <- tpm_122_Lumbar_noN_CV$Genes
bk_data <- tpm_122_Lumbar_noN_CV[,c(-1,-ncol(tpm_122_Lumbar_noN_CV))]
rownames(bk_data) <- Genes 
```

## 1.3 Constract SingleCellExperiment
```{r}
sce <- SingleCellExperiment(
  assays = list(
    counts = as.matrix(bk_data),
    logcounts = log2(as.matrix(bk_data)+1)
  )
)
rowData(sce)$feature_symbol <- rownames(sce)
```

# 2 Analysis
## 2.1 Estimate k
```{r}
# sce4 <- sc3(sce, ks = 3, biology = TRUE,gene_filter = FALSE) #这一步会运行很久
# data3 <- as.data.frame(rowData(sce3))
sce <- sc3(sce, ks = 3:4, biology = TRUE,
           # pct_dropout_min = 10,
           # pct_dropout_max = 90,
           gene_filter = F) #这一步会运行很久

sce<-sc3_estimate_k(sce) #估计k值
ks<-metadata(sce)$sc3$k_estimation #调用k值
ks
```
## 2.2 PCA
```{r}
set.seed(1)
sce <- runPCA(sce)
plotPCA(sce,
        #size_by = "sc3_3_log2_outlier_score",
        colour_by = "sc3_3_clusters"
        )
```
## 2.2 Consistent matrix
```{r}
#计算一致性矩阵
sc3_plot_consensus(
    sce, k = 3,
    show_pdata = c(
        #"cell_type", 
        #"log10_total_features",
        "sc3_3_log2_outlier_score",
        "sc3_2_clusters",
        "sc3_3_clusters",
        "sc3_4_clusters",
        "sc3_5_clusters",
        "sc3_6_clusters"

    )
)
```

```{r}
res <- as.data.frame(colnames(L_bk_noB))
for(i in seq(2,5)){
  a <- sc3_plot_consensus(sce, k = i)
  res <- cbind(res,a[["tree_row"]][["order"]])
}
colnames(res) <- c('cluster','n=2','n=3','n=4','n=5')
openxlsx::write.xlsx(res,
                     '/linp/linp/IDD_classification/105Samples/Protocol_test/SC3_All_gene/SC3_10.90/Hierachy_order_2-5.xlsx',
                     overwrite = FALSE)
```

## 2.3 ROC marker genes
```{r}
sc3_plot_markers(sce, k = 3)
```
## 2.4 DE
```{r}
#sc3_plot_de_genes(sce, k = 3)
DEG <- as.data.frame(rowData(sce))
Row <- as.data.frame(colData(sce))
DEG[DEG$sc3_gene_filter,] %>% nrow()

# DEG.3 <- DEG[,c(1,6,7,8)]
# DEG.3 <- DEG.3[DEG.3$sc3_3_markers_padj < 0.05,]
# DEG.3[!is.na(DEG.3$sc3_3_markers_clusts),] %>% nrow()
# 
# DEG.4 <- DEG[,c(1,9,10,11)]
# DEG.4 <- DEG.4[DEG.4$sc3_4_markers_padj < 0.05,]
# DEG.4 <- DEG.4[!is.na(DEG.4$sc3_4_markers_clusts),]
```

## 2.5 Silhouette plot
```{r}
# Silhouette_Coef=6
color <- c("red", "green", "blue", "purple","orange","yellow")
for(i in seq(3,4)){
  plot(sce@metadata[["sc3"]][["consensus"]][[i-2]][["silhouette"]], 
     col = color[1:i])
}
# sc3_plot_silhouette(sce, k = 5)
# plot(sce@metadata[["sc3"]][["consensus"]][["4"]][["silhouette"]],
#      col = c("red", "green", "blue", "purple"))
# fviz_silhouette(sce@metadata[["sc3"]][["consensus"]][["3"]][["silhouette"]])

#data.frame(colnames(bk_data),matrix(sce@metadata[["sc3"]][["consensus"]][[3]][["silhouette"]])[221:330]) %>% View()
```
### Save outlier
```{r}
Row <- as.data.frame(colData(sce))
outpath <- "/linp/linp/IDD_classification/Formal Analysis/3 Clustering/122Lumbar/New_5/122Lumbar_SC3_CV3500/outlier="
for(i in seq(3,4)){
  a <- data.frame(colnames(bk_data),matrix(sce@metadata[["sc3"]][["consensus"]][[i-1]][["silhouette"]])[245:366])
  colnames(a) <- c('sample','silhouette_coef')
  a$outlier_score <- Row[,4+i]
  openxlsx::write.xlsx(a,
                       paste(outpath,i,'.xlsx',sep = ""),
                       overwrite = F)
}

```

### Silhouette trend plot
```{r}
sc = c()
k = 0 
for(i in seq(2,6)){
  sc <- c(sc,
          mean(matrix(sce@metadata[["sc3"]][["consensus"]][[i-2]][["silhouette"]])
               [245:364])
          )
}

pic2 <- as.data.frame(sc)
colnames(pic2) <- c('Sil_Coef')
pic2$k <- seq(2,6)
# pic2$method <- "SC3"
# pic2$HVGs <- "10-90"
```

```{r}
# Sil_Coef_trend_2-6
ggplot(data = pic2, 
       mapping = aes(x = k, y = Sil_Coef 
                     # colour = method, 
                     # linetype = HVGs, 
                     # shape = HVGs, fill = HVGs
                     )) + 
  geom_line()+geom_point()+
  geom_text_repel(
    aes(x = k, y = Sil_Coef, label = round(Sil_Coef,3)),
    #size = 2
                  ) +
  #ylim(0,1)+
  xlab("Number of clusters") + ylab("Value")+labs(title = "SC3_CV11325")+
  theme_bw() + #去除背景色
  theme(plot.title = element_text(hjust = 0.5))+
  theme(panel.grid =element_blank()) #去除网格线

ggsave("/linp/linp/IDD_classification/Formal Analysis/3 Clustering/122Lumbar/Threshold=5/122Lumbar_SC3_CV11325/Silhouette_Coef.pdf",
       plot = last_plot(),
       device = cairo_pdf,width =6, height =4,dpi = 300)
```

# 3 Data saving
```{r}
outpath <- "/linp/linp/IDD_classification/Formal Analysis/3 Clustering/122Lumbar/Threshold=5/122Lumbar_SC3_CV11325/Group="
for(i in seq(2,6)){
  data <- as.data.frame(colData(sce))
  group <- as.data.frame(data[,i-1])
  group$samples <- colnames(bk_data)

  G_data <- group
  colnames(G_data) <- c('cluster','samples')
  openxlsx::write.xlsx(G_data,
                     paste(outpath,i,'.xlsx',sep = ""),
                     overwrite = FALSE)
}
```

# 4 Cluster stability
## 1 Circulation
```{r}
Sil_data <- data.frame(HVGs=c(),Group=c(),Sil_Score=c(),Method=c())
for(gene_no in seq(2500,11000,500)){
  target <- tpm_122_Lumbar_noN_CV[order(tpm_122_Lumbar_noN_CV$CV,decreasing = T),][1:gene_no,]
  Genes <- target$Genes
  bk_data <- target[,c(-1,-ncol(target))]
  rownames(bk_data) <- Genes 
  
  sce <- SingleCellExperiment(
  assays = list(
    counts = as.matrix(bk_data),
    logcounts = log2(as.matrix(bk_data)+1)
    )
  )
  rowData(sce)$feature_symbol <- rownames(sce)
  
  sce <- sc3(sce, ks = 2:8, biology = TRUE,
           gene_filter = F)
  
  for(Group_no in seq(2,8)){
    sc <- mean(matrix(sce@metadata[["sc3"]][["consensus"]][[Group_no-1]][["silhouette"]])[245:364])
    pic <- data.frame(HVGs=gene_no , Group=Group_no , Sil_Score=sc , Method="SC3")
    Sil_data <- rbind(Sil_data,pic)
  }
  
  # outpath <- "/linp/linp/IDD_classification/Formal Analysis/3 Clustering/122Lumbar/New_5/Cluster_stability/Gene"
  # for(i in seq(3,4)){
  #   data <- as.data.frame(colData(sce))
  #   group <- as.data.frame(data[,i-2])
  #   group$samples <- colnames(bk_data)
  # 
  #   G_data <- group
  #   colnames(G_data) <- c('cluster','samples')
  #   openxlsx::write.xlsx(G_data,
  #                    paste(outpath,gene_no,'Group=',i,'.xlsx',sep = ""),
  #                    overwrite = FALSE)
  #}
}
openxlsx::write.xlsx(Sil_data,
                     "/linp/linp/IDD_classification/Formal Analysis/3 Clustering/122Lumbar/New_5/SC3_stability/Sil_data_2500-11000(2~8).xlsx",
                     overwrite = F)
```

## 2 Picture
```{r}
Sil_data <- read.xlsx("/linp/linp/IDD_classification/Formal Analysis/3 Clustering/122Lumbar/New_5/SC3_stability/Sil_data_2500-11000(2~8).xlsx")
Sil_data$Group <- as.character(Sil_data$Group)
ggplot(data = Sil_data, 
       mapping = aes(x = HVGs, y = Sil_Score ,
                     colour = Group, 
                     #linetype = Group, 
                     # shape = HVGs, fill = HVGs
                     )) + 
  geom_line()+geom_point()+
  # geom_text_repel(
  #   aes(x = HVGs, y = Sil_Score, label = round(Sil_Score,3)),
  #   size = 3
  #                 ) +
  #ylim(0,1)+
  xlab("HVGs numbers") + ylab("Sil_Score")+labs(title = "SC3 HVGs 2500~11000")+
  theme_bw() + #去除背景色
  theme(plot.title = element_text(hjust = 0.5))+
  theme(panel.grid =element_blank()) #去除网格线

# ggsave("/linp/linp/IDD_classification/Formal Analysis/3 Clustering/122Lumbar/New_5/SC3_stability/Sil_data_2500-11000(2~8)_noNum.pdf",
#        plot = last_plot(),
#        device = cairo_pdf,width =6, height =4,dpi = 300)
```

### CV8000
```{r}
library(ggrepel)
# Sil_data <- read.xlsx("/linp/linp/IDD_classification/Formal Analysis/3 Clustering/122Lumbar/New_5/SC3_stability/Sil_data_2500-11000(2~8).xlsx")
Sil_data$Group <- as.character(Sil_data$Group)
Sil_data$Sil_Score <- as.numeric(Sil_data$Sil_Score)
Sil_data <- Sil_data[Sil_data$HVGs == 8000,]

ggplot(data = Sil_data, 
       mapping = aes(x = Group, y = Sil_Score,group = HVGs,
                     #colour = Group, 
                     #linetype = Group, 
                     # shape = HVGs, fill = HVGs
                     )) + 
  geom_line()+
  geom_point()+
  geom_text_repel(
    aes(x = Group, y = Sil_Score, label = round(Sil_Score,3)),
    size = 3
                  ) +
  ylim(0.4,1)+
  xlab("Groups") + ylab("Sil_Score")+labs(title = "SC3 HVGs8000")+
  theme_bw() + #去除背景色
  theme(plot.title = element_text(hjust = 0.5))+
  theme(panel.grid =element_blank()) #去除网格线

# ggsave("/linp/linp/IDD_classification/Formal Analysis/3 Clustering/122Lumbar/New_5/SC3_stability/Sil_data_8000(2~8).pdf",
#        plot = last_plot(),
#        device = cairo_pdf,width =6, height =4,dpi = 300)

```

# 5 PCA 
```{r}
tpm_122_Lumbar_noN_CV <- read.xlsx("/linp/linp/IDD_classification/Formal Analysis/1 Primary matrix/TPM_122_Lumbar_noN_CV_5.xlsx")
Group <- read.xlsx("/linp/linp/IDD_classification/Formal Analysis/3 Clustering/122Lumbar/New_5/SC3_stability/Gene8000Group=4.xlsx")
target <- tpm_122_Lumbar_noN_CV[order(tpm_122_Lumbar_noN_CV$CV,decreasing = T),][1:8000,]
target <- target[,c(-1,-ncol(target))]
t_target <- t(target)
```

```{r}
set.seed(1)
#t_scale[ , which(apply(t_scale, 2, var) != 0)]
pca1 <- prcomp(t_target[ , which(apply(t_target, 2, var) != 0)],
               center = TRUE,
               scale. = TRUE)
df1 <- pca1$x # 提取PC score
df1 <- as.data.frame(df1) # 注意：如果不转成数据框形式后续绘图时会报错
```

```{r}
#group <- c(rep('batch1',80),rep('batch2',40))
library(ggalt)
library(ggrepel)
summ1 <- summary(pca1)
xlab1 <- paste0("PC4(",round(summ1$importance[2,4]*100,2),"%)")
ylab1 <- paste0("PC10(",round(summ1$importance[2,10]*100,2),"%)")

ggplot(data = df1,aes(x = PC4,y = PC10,
                      color = as.character(
                        Group$cluster
                      )
                      ))+
  geom_point(size = 2)+
  labs(x = xlab1,y = ylab1,color = "Group",title = "New5 SC3 CV8000 n=4")+
  guides(fill = "none")+   
  #geom_text_repel(aes(PC1, PC2, label = rownames(df1)))+
  # stat_ellipse(
  #   # aes(fill=as.character(                  #加置信椭圆
  #   #                     group$cluster
  #   #                   )),
  #              level = 0.95,
  #              type = "norm",
  #              geom ="polygon",
  #              alpha=0.1)+
  geom_encircle(
    aes(fill=as.character(
                        Group$cluster
                      )),
                expand=0,spread=0.5,s_shape=1,size=3,linetype = 1,alpha=0.2)+
  theme_bw() + #去除背景色
  theme(plot.title = element_text(size=12,hjust=0.5))+  #标题居中
  theme(panel.grid =element_blank()) #去除网格线
  # guides(fill = "none")+
  # theme_bw()+
  # scale_fill_manual(values = c("purple","orange","pink"))+
  # scale_colour_manual(values = c("purple","orange","pink"))
  # theme(plot.title = element_text(hjust = 0.5,size = 15),
  #       axis.text = element_text(size = 11),axis.title = element_text(size = 13),
  #       legend.text = element_text(size = 11),legend.title = element_text(size = 13),
  #       plot.margin = unit(c(0.4,0.4,0.4,0.4),'cm'))

# ggsave("/linp/linp/IDD_classification/Formal Analysis/3 Clustering/122Lumbar/New_5/NMF_PCA/New5 NMF CV3000 n=4.pdf",
#          plot = last_plot(),
#          device = cairo_pdf,width =8, height =6,dpi = 300)
```
### 1 3D
```{r}
library(scatterplot3d)
colors <- c("red","green","blue","purple")
colors <- colors[as.numeric(as.factor(Group$cluster))]
scatterplot3d(df1[,c(4,1,2)],
                     pch = 16,       # 点形状
                     color=colors,   # 点颜色
                     cex.symbols = 1 # 点大小
                     )
```

# 6 t-SNE
## 1 Library
```{r}
library(pheatmap)
library(Rtsne)
```

## 2 Preprocess
```{r}
gene_no <- 8000
#tpm_122_Lumbar_noN_CV <- read.xlsx("/linp/linp/IDD_classification/Formal Analysis/1 Primary matrix/TPM_122_Lumbar_noN_CV_5.xlsx")
target <- tpm_122_Lumbar_noN_CV[order(tpm_122_Lumbar_noN_CV$CV,decreasing = T),][1:gene_no,]
target <- target[,c(-1,-ncol(target))]
t_target <- t(target)
```

```{r}
tsne_out <- Rtsne(t_target[ , which(apply(t_target, 2, var) != 0)],
                  pca=TRUE,perplexity=10,theta=0.0)
# 获取tSNE的坐标值
#str(tsne_out)
# 其中在Y中存储了画图坐标
tsnes=tsne_out$Y
colnames(tsnes) <- c("tSNE1", "tSNE2") #为坐标添加列名
# 在此基础上添加颜色分组信息，首先还是将tsnes这个矩阵变成数据框，然后增加一列group信息，最后映射在geom_point中
tsnes=as.data.frame(tsnes)
```

## 2.1 Visualization
```{r}
inpath <- "/linp/linp/IDD_classification/Formal Analysis/3 Clustering/122Lumbar/New_5/SC3_stability/Gene"
Group <- read.xlsx(paste(inpath,gene_no,"Group=4.xlsx",sep=""))
```

```{r,fig.width=6,fig.height=4}
ggplot(tsnes, aes(x = tSNE1, y = tSNE2))+ 
  geom_point(aes(col=as.character(
    Group$cluster
    )), size = 2)+
  labs(colour = "Cluster")+
  #geom_text_repel(aes(tSNE1, tSNE2, label = rownames(df1)))+
  geom_encircle(
    aes(fill=as.character(
                        Group$cluster
                      )),
                expand=0,spread=0.5,s_shape=1,size=3,linetype = 1,alpha=0.2)+
  theme_bw() + #去除背景色
  theme(panel.grid =element_blank(), #去除网格线
        panel.border = element_blank())+  #去边框
  theme(axis.line = element_line(arrow = arrow(length = unit(0.5, 'cm'))))  #坐标轴尾端为箭头

# kmeans(t_scale, centers = 5, nstart = 25)$cluster  #kmean
# cutree(produc_hc,k=5)  #heirachy
# results[[5]][["consensusClass"]] #consensus
# group$group  #NMF
```
