==> 6.3 Scissor.Rmd <==

# 1 Preparation
## 1.1 Library
```{r}
library(Scissor)
library(Seurat)
library(preprocessCore)
library(ggplot2)
```

## 1.2 AS_BR data and test
### 1 Primary Process
```{r, fig.height=5,fig.width=5.5}
color2 <- c(rgb(214/256,39/256,40/256),  # Chond1
            rgb(148/256,103/256,189/256),# Chond2
            rgb(140/256,86/256,75/256),  # Chond3
            rgb(227/256,119/256,194/256),# NC
            rgb(31/256,119/256,180/256),  # NPPC
            rgb(255/256,126/256,14/256),  # Stroma
            rgb(188/256,189/256,34/256), # EC
            rgb(44/256,160/256,44/256),  # Pericyte
            rgb(23/256,190/256,207/256), # Blood
            rgb(197/256,168/256,206/256), # Neutrophils
            rgb(239/256,132/256,118/256), # GMPs
            rgb(249/256,219/256,149/256), # Macrophage
            rgb(140/256,205/256,191/256)  # T cell
            ) 
AS_BR <- readRDS("/hejian1/Scissors/AS_BR/Primary_data/Merge-20231012-2.rds")
DimPlot(AS_BR, reduction = "umap",group.by = 'furcluster',
        cols = color2
        )
AS_BR<- Seurat_preprocessing(AS_BR@assays[["RNA"]]@counts, verbose = F)  # Scissor 一定需要这一步处理
AS_BR.2 <- readRDS("/hejian1/Scissors/AS_BR/Primary_data/Merge-20231012-2.rds")
AS_BR@meta.data[["furcluster"]] <- AS_BR.2@meta.data[["furcluster"]]
AS_BR@meta.data[["source"]] <- AS_BR.2@meta.data[["source"]]
AS_BR@reductions[["umap"]] <- AS_BR.2@reductions[["umap"]]
#saveRDS(AS_BR,"/hejian1/Scissors/AS_BR/Primary_data/Merge-20231013.rds")

AS <- subset(AS_BR,(source == c("AS")))
BR <- subset(AS_BR,(source == c("BR")))

AS.2 <- Seurat_preprocessing(AS@assays[["RNA"]]@counts, verbose = F)
AS.2@meta.data[["furcluster"]] <- AS@meta.data[["furcluster"]]
AS.2@meta.data[["source"]] <- AS@meta.data[["source"]]
AS.2@reductions[["umap"]] <- AS@reductions[["umap"]]

BR.2 <- Seurat_preprocessing(BR@assays[["RNA"]]@counts, verbose = F)
BR.2@meta.data[["furcluster"]] <- BR@meta.data[["furcluster"]]
BR.2@meta.data[["source"]] <- BR@meta.data[["source"]]
BR.2@reductions[["umap"]] <- BR@reductions[["umap"]]

saveRDS(AS.2,"/hejian1/Scissors/AS_BR/Primary_data/AS-20231013.rds")
saveRDS(BR.2,"/hejian1/Scissors/AS_BR/Primary_data/BR-20231013.rds")
```

### 2 Only data read in
```{r,fig.height=5,fig.width=5.5}
color2 <- c(rgb(214/256,39/256,40/256),  # Chond1
            rgb(148/256,103/256,189/256),# Chond2
            rgb(140/256,86/256,75/256),  # Chond3
            rgb(227/256,119/256,194/256),# NC
            rgb(31/256,119/256,180/256),  # NPPC
            rgb(255/256,126/256,14/256),  # Stroma
            rgb(188/256,189/256,34/256), # EC
            rgb(44/256,160/256,44/256),  # Pericyte
            rgb(23/256,190/256,207/256), # Blood
            rgb(197/256,168/256,206/256), # Neutrophils
            rgb(239/256,132/256,118/256), # GMPs
            rgb(249/256,219/256,149/256), # Macrophage
            rgb(140/256,205/256,191/256)  # T cell
            ) 
AS_BR <- readRDS("/hejian1/Scissors/AS_BR/Primary_data/Merge-20231013.rds")
DimPlot(AS_BR, reduction = "umap",group.by = 'furcluster',
        cols = color2
        )
```

```{r}
color2 <- c(rgb(214/256,39/256,40/256),  # Chond1
            rgb(148/256,103/256,189/256),# Chond2
            rgb(140/256,86/256,75/256),  # Chond3
            rgb(227/256,119/256,194/256),# NC
            rgb(31/256,119/256,180/256),  # NPPC
            rgb(255/256,126/256,14/256),  # Stroma
            rgb(188/256,189/256,34/256), # EC
            rgb(44/256,160/256,44/256),  # Pericyte
            rgb(23/256,190/256,207/256), # Blood
            rgb(197/256,168/256,206/256), # Neutrophils
            rgb(239/256,132/256,118/256), # GMPs
            rgb(249/256,219/256,149/256), # Macrophage
            rgb(140/256,205/256,191/256)  # T cell
            ) 
AS <- readRDS("/hejian1/Scissors/AS_BR/Primary_data/AS-20231013.rds")
BR <- readRDS("/hejian1/Scissors/AS_BR/Primary_data/BR-20231013.rds")
```

```{r}
Feature <- c("PECAM1","VWF",'CAV1',
             "ACTA2","MYH11",
             "ACAN","COL2A1",
             "KRT8","TBXT",
             'PROCR','PDGFRA',"CTSK",'PRG4',
             'FBLN1',"COL6A1","COL1A2","COL1A1",
             "LYZ","CD14","CSF3R",
             "CD3D","CD8A",#"CD3E","RUNX3","TRAC",
             "IBSP","SPP1","SP7","RUNX2"
             #"NGFR","SOX2","NRXN1"
             )
#VlnPlot(all_unfra, features = Feature,pt.size = 0,ncol = 2)
DotPlot(BR_data2, features = Feature, 
        #cluster.idents = TRUE,
        #group.by = 'RNA_snn_res.0.3',
        cols ="RdBu") + RotatedAxis()
```
### 3 scRNA analysis
```{r}
# AS <- subset(AS_BR,(source == "AS"))
# #Idents(AS_BR) <- AS_BR@meta.data[["furcluster"]]
# Idents(AS) <- AS@meta.data[["furcluster"]]
#VlnPlot(AS_BR, features = c("HIF1A"),pt.size = 0)
Blood_featrue <- c("CD14","CD68","CD86","CD163",
                   "ITGAL","ITGAM","ITGAX","CST3","LYZ")
DotPlot(AS, features =  Blood_featrue,
          #c("HIF1A","PECAM1","PTPRC","ITGAM","CD68"),
        #split.by = "source",
        cols = "RdBu" ) + 
  RotatedAxis()
```

#### 1 Diff Mac Markers
```{r}
Features <- c("IFNG","IL6","TNF",  # M1 Mac
              "IL10","TGFB1","ARG1", # M2 Mac
              "TMEM119", # 小胶质
              "TIMD4", # Kuffer cell
              "TRAP","CTSK", # Osteoclast
              "CD86","ITGAL","ITGAM","ITGAX"
              ) 
DotPlot(AS, features =  Features,
        #split.by = "source",
        cols = "RdBu" ) + 
  RotatedAxis()
```

## 1.3 Bulk data and test
```{r}
library(openxlsx)
library(readxl)

Bulk_data = read_xlsx("/hejian1/Scissors/122Lumbar/Primary data/TPM_122_Lumbar.xlsx")
genes <- Bulk_data$Genes
Bulk_data <- Bulk_data[,-1]
rownames(Bulk_data) <- genes

table(rowSums(Bulk_data) == 0)
```

# 2 Scissors
## 2.1 Test data
### 2.1.1 Test1
```{r}
# library(preprocessCore)
# library(oligo)

location <- "https://xialab.s3-us-west-2.amazonaws.com/Duanchen/Scissor_data/"
load(url(paste0(location, 'scRNA-seq.RData')))
dim(sc_dataset)
sc_dataset <- Seurat_preprocessing(sc_dataset, verbose = F)
class(sc_dataset)
names(sc_dataset)

load(url(paste0(location, 'TCGA_LUAD_exp1.RData')))
load(url(paste0(location, 'TCGA_LUAD_survival.RData')))
dim(bulk_dataset)
head(bulk_survival)
all(colnames(bulk_dataset) == bulk_survival$TCGA_patient_barcode)

phenotype <- bulk_survival[,2:3]
colnames(phenotype) <- c("time", "status")
head(phenotype)

#preprocessCore::normalize.quantiles(bulk_dataset,copy = T)
infos1 <- Scissor(bulk_dataset, sc_dataset, phenotype, alpha = 0.05, 
                 family = "cox", Save_file = 'Scissor_LUAD_survival.RData')
```

### 2.1.2 Test2
```{r}
load(url(paste0(location, 'TCGA_LUAD_TP53_mutation.RData')))
load(url(paste0(location, 'TCGA_LUAD_exp2.RData')))
table(TP53_mutation)
phenotype <- TP53_mutation
tag <- c('wild-type', 'TP53 mutant')

infos4 <- Scissor(bulk_dataset = bulk_dataset, sc_dataset = sc_dataset, 
                   phenotype = TP53_mutation, tag = tag, alpha = 0.5,
                   family = "binomial", 
                   #Save_file = "Scissor_LUAD_TP53_mutation.RData"
                   )
```

## 2.2 Analysis
```{r}
a <- c(rep(0,3),rep(1,5))
b <- as.data.frame(a)
Segment <- t(a)
Segment <- as.numeric(Segment)
names(Segment) <- colnames(Bulk_data)

tag <- c('Cluster1', 'Cluster2')
Bulk_data_2 <- as.matrix(Bulk_data)

scissor <- Scissor(bulk_dataset = Bulk_data_2, sc_dataset = BR_NP, 
                   phenotype = Segment, tag = tag, alpha = 0.5,
                   family = "binomial", 
                   Save_file = "/hejian1/Scissors/Res_data/Scissor_8Sam.RData"
                   )
```

```{r}
Res <- data.frame(c(scissor$Scissor_pos,rep(NA,515-179)),scissor$Scissor_neg)
colnames(Res) <- c("Scissor_pos","Scissor_neg")
openxlsx::write.xlsx(Res,
                     "/hejian1/Scissors/Res_data/Scissor_8Sam.xlsx",
                     overwrite = F)
```

```{r,fig.height=4,fig.width=5}
Res <- read.xlsx("/hejian1/Scissors/Res_data/Scissor_8Sam.xlsx")
Scissor_select <- rep(0, ncol(BR_NP))
names(Scissor_select) <- colnames(BR_NP)
Scissor_select[Res$Scissor_pos] <- 1
Scissor_select[Res$Scissor_neg] <- 2

BR_NP <- AddMetaData(BR_NP, metadata = Scissor_select, col.name = "scissor")
DimPlot(BR_NP, reduction = 'tsne', group.by = 'scissor', 
        cols = c('grey','indianred1','royalblue'), 
        pt.size = 1.2, order = c(2,1))
```

# 3 122 Lumbar
## 1 Data
```{r}
# 和1.3 一致
library(openxlsx)
library(readxl)

Bulk_data = read_xlsx("/hejian1/Scissors/122Lumbar/Primary data/TPM_122_Lumbar.xlsx")
genes <- Bulk_data$Genes
Bulk_data <- Bulk_data[,-1]
rownames(Bulk_data) <- genes

table(rowSums(Bulk_data) == 0)
```

## 2 C3
### 1 Analysis
```{r}
group <- read_xlsx("/hejian1/Scissors/122Lumbar/Primary data/Gene8000Group=4.xlsx")
group[group$cluster != '3',]$cluster <- 'rest'
group[group$cluster == '3',]$cluster <- '1'
group[group$cluster == 'rest',]$cluster <- '0'

Segment <- as.numeric(group$cluster)  #pos：1，neg：0
names(Segment) <- colnames(Bulk_data)

tag <- c('Rest','Cluster3')  # 对应：0,1
Bulk_data_2 <- as.matrix(Bulk_data)

scissor <- Scissor(bulk_dataset = Bulk_data_2, sc_dataset = BR, 
                   phenotype = Segment, tag = tag, alpha = 0.5,
                   family = "binomial", 
                   Save_file = "/hejian1/Scissors/AS_BR/Res_data/RData/Scissor_122Lum_BR_C3.RData"
                   )
```

```{r}
Res <- data.frame(c(scissor$Scissor_pos),c(scissor$Scissor_neg,rep(NA,1450-823)))
colnames(Res) <- c("Scissor_pos","Scissor_neg")
openxlsx::write.xlsx(Res,
                     "/hejian1/Scissors/AS_BR/Res_data/xlsx/Scissor_122Lum_BR_C3.xlsx",
                     overwrite = F)
```
### 2 Visualization
```{r,fig.height=4,fig.width=5}
Res <- read.xlsx("/hejian1/Scissors/AS_BR/Res_data/xlsx/Scissor_122Lum_BR_C3.xlsx")
Scissor_select <- rep(0, ncol(BR))
names(Scissor_select) <- colnames(BR)
Scissor_select[Res$Scissor_pos] <- 1
Scissor_select[Res$Scissor_neg] <- 2

BR <- AddMetaData(BR, metadata = Scissor_select, col.name = "scissor")
DimPlot(BR, reduction = 'umap', group.by = 'scissor', 
        cols = c('grey','indianred1','royalblue'), 
        pt.size = 0.1, order = c(2,1))
ggsave(paste("/hejian1/Scissors/AS_BR/Res_data/Figures/BR_UMAP_scissor_C3.pdf"),
       plot = last_plot(),
       device = cairo_pdf,width =5, height =5,dpi = 300)
```

```{r}
pos <- subset(BR,(scissor == "1"))
neg <- subset(BR,(scissor == "2"))
```

```{r}
pos_pie <- table(pos@meta.data[["furcluster"]])
neg_pie <- table(neg@meta.data[["furcluster"]])
pos_pie
neg_pie
```

```{r}
mycolor <- c(rgb(31/256,119/256,180/256),
                 rgb(255/256,126/256,14/256),
                 rgb(44/256,160/256,44/256),
                 rgb(214/256,39/256,40/256),
                 rgb(148/256,103/256,189/256),
                 rgb(140/256,86/256,75/256),
                 rgb(227/256,119/256,194/256),
                 rgb(188/256,189/256,34/256),
                 rgb(23/256,190/256,207/256))

Pie <- as.data.frame(pos_pie)
colnames(Pie) <- c('Cluster','Count')
Pie$Cluster <- factor(Pie$Cluster,
                          levels = Pie$Cluster)
Pie$Percent <- Pie$Count/sum(Pie$Count)

pie(Pie$Percent, labels = with(Pie, paste0(round(Percent, 2) * 100, "%")), 
    col = mycolor)

ggplot(Pie,aes(x = factor(1), y = Percent, fill = Cluster)) +
  geom_col(colour = "white")+
  coord_polar(theta = "y")+
  scale_fill_manual(values=color2)+
  geom_text(aes(label = paste0(round(Percent * 100, 2), "%")), 
            position = position_fill(vjust = 0.5)) +
  theme(
    panel.background = element_blank(),
    axis.title = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank()
  )
ggsave(paste("/hejian1/Scissors/AS_BR/Res_data/Figures/BR_Pie_C3.pos.pdf"),
       plot = last_plot(),
       device = cairo_pdf,width =5, height =5,dpi = 300)
```

## 3 C2
### 1 Analysis
```{r}
group <- read_xlsx("/hejian1/Scissors/122Lumbar/Primary data/Gene8000Group=4.xlsx")
group[group$cluster != '2',]$cluster <- 'rest'
group[group$cluster == '2',]$cluster <- '1'
group[group$cluster == 'rest',]$cluster <- '0'

Segment <- as.numeric(group$cluster)  #pos：1，neg：0
names(Segment) <- colnames(Bulk_data)

tag <- c('Rest','Cluster2')  # 对应：0,1
Bulk_data_2 <- as.matrix(Bulk_data)

scissor <- Scissor(bulk_dataset = Bulk_data_2, sc_dataset = BR, 
                   phenotype = Segment, tag = tag, alpha = 0.5,
                   family = "binomial", 
                   Save_file = "/hejian1/Scissors/AS_BR/Res_data/RData/Scissor_122Lum_BR_C2.RData"
                   )
```

```{r}
Res <- data.frame(c(scissor$Scissor_pos),c(scissor$Scissor_neg,rep(NA,1268-1057)))
colnames(Res) <- c("Scissor_pos","Scissor_neg")
openxlsx::write.xlsx(Res,
                     "/hejian1/Scissors/AS_BR/Res_data/xlsx/Scissor_122Lum_BR_C2.xlsx",
                     overwrite = F)
```

### 2 Visualization
```{r,fig.height=4,fig.width=5}
Res <- read.xlsx("/hejian1/Scissors/AS_BR/Res_data/xlsx/Scissor_122Lum_BR_C2.xlsx")
Scissor_select <- rep(0, ncol(BR))
names(Scissor_select) <- colnames(BR)
Scissor_select[Res$Scissor_pos] <- 1
Scissor_select[Res$Scissor_neg] <- 2

BR <- AddMetaData(BR, metadata = Scissor_select, col.name = "scissor")
DimPlot(BR, reduction = 'umap', group.by = 'scissor', 
        cols = c('grey','indianred1','royalblue'), 
        pt.size = 0.1, order = c(2,1))
ggsave(paste("/hejian1/Scissors/AS_BR/Res_data/Figures/BR_UMAP_scissor_C2.pdf"),
       plot = last_plot(),
       device = cairo_pdf,width =5, height =5,dpi = 300)
```

```{r}
pos <- subset(BR,(scissor == "1"))
neg <- subset(BR,(scissor == "2"))
```

```{r}
pos_pie <- table(pos@meta.data[["furcluster"]])
neg_pie <- table(neg@meta.data[["furcluster"]])
pos_pie
neg_pie
```

```{r}
mycolor <- c(rgb(31/256,119/256,180/256),
                 rgb(255/256,126/256,14/256),
                 rgb(44/256,160/256,44/256),
                 rgb(214/256,39/256,40/256),
                 rgb(148/256,103/256,189/256),
                 rgb(140/256,86/256,75/256),
                 rgb(227/256,119/256,194/256),
                 rgb(188/256,189/256,34/256),
                 rgb(23/256,190/256,207/256))

Pie <- as.data.frame(pos_pie)
colnames(Pie) <- c('Cluster','Count')
Pie$Cluster <- factor(Pie$Cluster,
                          levels = Pie$Cluster)
Pie$Percent <- Pie$Count/sum(Pie$Count)

# pie(Pie$Percent, labels = with(Pie, paste0(round(Percent, 2) * 100, "%")), 
#     col = mycolor)

ggplot(Pie,aes(x = factor(1), y = Percent, fill = Cluster)) +
  geom_col(colour = "white")+
  coord_polar(theta = "y")+
  scale_fill_manual(values=color2)+
  geom_text(aes(label = paste0(round(Percent * 100, 2), "%")), 
            position = position_fill(vjust = 0.5)) +
  theme(
    panel.background = element_blank(),
    axis.title = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank()
  )
ggsave(paste("/hejian1/Scissors/AS_BR/Res_data/Figures/BR_Pie_C2.pos.pdf"),
       plot = last_plot(),
       device = cairo_pdf,width =5, height =5,dpi = 300)

```

## 4 C1
### 1 Analysis
```{r}
group <- read_xlsx("/hejian1/Scissors/122Lumbar/Primary data/Gene8000Group=4.xlsx")
group[group$cluster != '1',]$cluster <- 'rest'
group[group$cluster == '1',]$cluster <- '1'
group[group$cluster == 'rest',]$cluster <- '0'

Segment <- as.numeric(group$cluster)  #pos：1，neg：0
names(Segment) <- colnames(Bulk_data)

tag <- c('Rest', 'Cluster1')  # 对应：0,1
Bulk_data_2 <- as.matrix(Bulk_data)

scissor <- Scissor(bulk_dataset = Bulk_data_2, sc_dataset = BR, 
                   phenotype = Segment, tag = tag, alpha = 0.5,
                   family = "binomial", 
                   Save_file = "/hejian1/Scissors/AS_BR/Res_data/RData/Scissor_122Lum_BR_C1.RData"
                   )
```

```{r}
#load("/hejian1/Scissors/AS_BR/Res_data/RData/Scissor_122Lum_BR_C1.RData")
Res <- data.frame(c(scissor$Scissor_pos,rep(NA,591-566)),c(scissor$Scissor_neg))
colnames(Res) <- c("Scissor_pos","Scissor_neg")
openxlsx::write.xlsx(Res,
                     "/hejian1/Scissors/AS_BR/Res_data/xlsx/Scissor_122Lum_C1.xlsx",
                     overwrite = F)
```
### 2 Visualization
```{r,fig.height=4,fig.width=5}
Res <- read.xlsx("/hejian1/Scissors/AS_BR/Res_data/xlsx/Scissor_122Lum_BR_C1.xlsx")
Scissor_select <- rep(0, ncol(BR))
names(Scissor_select) <- colnames(BR)
Scissor_select[Res$Scissor_pos] <- 1
Scissor_select[Res$Scissor_neg] <- 2

BR <- AddMetaData(BR, metadata = Scissor_select, col.name = "scissor")
DimPlot(BR, reduction = 'umap', group.by = 'scissor', 
        cols = c('grey','indianred1','royalblue'), 
        pt.size = 0.1, order = c(2,1))
ggsave(paste("/hejian1/Scissors/AS_BR/Res_data/Figures/BR_UMAP_scissor_C1.pdf"),
       plot = last_plot(),
       device = cairo_pdf,width =5, height =5,dpi = 300)
```

```{r}
pos <- subset(BR,(scissor == "1"))
neg <- subset(BR,(scissor == "2"))
```

```{r}
pos_pie <- table(pos@meta.data[["furcluster"]])
neg_pie <- table(neg@meta.data[["furcluster"]])
pos_pie
neg_pie
```

```{r}
mycolor <- c(rgb(31/256,119/256,180/256),
                 rgb(255/256,126/256,14/256),
                 rgb(44/256,160/256,44/256),
                 rgb(214/256,39/256,40/256),
                 rgb(148/256,103/256,189/256),
                 rgb(140/256,86/256,75/256),
                 rgb(227/256,119/256,194/256),
                 rgb(188/256,189/256,34/256),
                 rgb(23/256,190/256,207/256))

Pie <- as.data.frame(neg_pie)
colnames(Pie) <- c('Cluster','Count')
Pie$Cluster <- factor(Pie$Cluster,
                          levels = Pie$Cluster)
Pie$Percent <- Pie$Count/sum(Pie$Count)

# pie(Pie$Percent, labels = with(Pie, paste0(round(Percent, 2) * 100, "%")), 
#     col = mycolor)

ggplot(Pie,aes(x = factor(1), y = Percent, fill = Cluster)) +
  geom_col(colour = "white")+
  coord_polar(theta = "y")+
  scale_fill_manual(values=color2)+
  geom_text(aes(label = paste0(round(Percent * 100, 2), "%")), 
            position = position_fill(vjust = 0.5)) +
  theme(
    panel.background = element_blank(),
    axis.title = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank()
  )

ggsave(paste("/hejian1/Scissors/AS_BR/Res_data/Figures/BR_Pie_C1.neg.pdf"),
       plot = last_plot(),
       device = cairo_pdf,width =5, height =5,dpi = 300)

```


## 5 C4
### 1 Analysis
```{r}
group <- read_xlsx("/hejian1/Scissors/122Lumbar/Primary data/Gene8000Group=4.xlsx")
group[group$cluster != '4',]$cluster <- 'rest'
group[group$cluster == '4',]$cluster <- '1'
group[group$cluster == 'rest',]$cluster <- '0'

Segment <- as.numeric(group$cluster)  #pos：1，neg：0
names(Segment) <- colnames(Bulk_data)

tag <- c('Rest','Cluster4' )  # 对应：0,1
Bulk_data_2 <- as.matrix(Bulk_data)

scissor <- Scissor(bulk_dataset = Bulk_data_2, sc_dataset = AS, 
                   phenotype = Segment, tag = tag, 
                   alpha = 0.01, cutoff = 0.3, # α越小，细胞越多
                   family = "binomial", 
                   Save_file = "/hejian1/Scissors/AS_BR/C4_20240226/Scissor_122Lum_AS_C4_0.01.RData"
                     #"/hejian1/Scissors/AS_BR/Res_data/RData/Scissor_122Lum_BR_C4.RData"
                   )
data.frame(c(scissor$Scissor_pos),c(scissor$Scissor_neg))
```

```{r}
Res <- data.frame(c(scissor$Scissor_pos),c(scissor$Scissor_neg,rep(NA,2792-2418))) # 
colnames(Res) <- c("Scissor_pos","Scissor_neg")
openxlsx::write.xlsx(Res,
                     "/hejian1/Scissors/AS_BR/C4_20240226/Scissor_122Lum_BR_C4_0.01.xlsx",
                     overwrite = F)
```
### 2 Visualization
```{r,fig.height=4,fig.width=5}
#Res <- read.xlsx("/hejian1/Scissors/AS_BR/C4_20240226/Scissor_122Lum_BR_C4.xlsx")
Scissor_select <- rep(0, ncol(AS))
names(Scissor_select) <- colnames(AS)
Scissor_select[Res$Scissor_pos] <- 1
Scissor_select[Res$Scissor_neg] <- 2

AS <- AddMetaData(AS, metadata = Scissor_select, col.name = "scissor")
DimPlot(AS, reduction = 'umap', group.by = 'scissor', 
        cols = c('grey','indianred1','royalblue'), 
        pt.size = 0.1, order = c(2,1))
ggsave(paste("/hejian1/Scissors/AS_BR/C4_20240226/BR_UMAP_scissor_C4_0.01.pdf"),
       plot = last_plot(),
       device = cairo_pdf,width =5, height =5,dpi = 300)
```

```{r}
pos <- subset(AS,(scissor == "1"))
neg <- subset(AS,(scissor == "2"))
```

```{r}
pos_pie <- table(pos@meta.data[["furcluster"]])
neg_pie <- table(neg@meta.data[["furcluster"]])
pos_pie
neg_pie
```

```{r}
mycolor <- c(rgb(31/256,119/256,180/256),
                 rgb(255/256,126/256,14/256),
                 rgb(44/256,160/256,44/256),
                 rgb(214/256,39/256,40/256),
                 rgb(148/256,103/256,189/256),
                 rgb(140/256,86/256,75/256),
                 rgb(227/256,119/256,194/256),
                 rgb(188/256,189/256,34/256),
                 rgb(23/256,190/256,207/256))

Pie <- as.data.frame(pos_pie)
colnames(Pie) <- c('Cluster','Count')
Pie$Cluster <- factor(Pie$Cluster,
                          levels = Pie$Cluster)
Pie$Percent <- Pie$Count/sum(Pie$Count)

# pie(Pie$Percent, labels = with(Pie, paste0(round(Percent, 2) * 100, "%")), 
#     col = mycolor)

ggplot(Pie,aes(x = factor(1), y = Percent, fill = Cluster)) +
  geom_col(colour = "white")+
  coord_polar(theta = "y")+
  scale_fill_manual(values=color2)+
  geom_text(aes(label = paste0(round(Percent * 100, 2), "%")), 
            position = position_fill(vjust = 0.5)) +
  theme(
    panel.background = element_blank(),
    axis.title = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank()
  )
ggsave(paste("/hejian1/Scissors/AS_BR/Res_data/Figures/BR_Pie_C4.pos.pdf"),
       plot = last_plot(),
       device = cairo_pdf,width =5, height =5,dpi = 300)

```

# 4 All pie
```{r}
all_pie <- table(BR@meta.data[["furcluster"]])
color2 <- c(rgb(214/256,39/256,40/256),  # Chond1
            rgb(148/256,103/256,189/256),# Chond2
            rgb(140/256,86/256,75/256),  # Chond3
            rgb(227/256,119/256,194/256),# NC
            rgb(31/256,119/256,180/256),  # NPPC
            rgb(255/256,126/256,14/256),  # Stroma
            rgb(188/256,189/256,34/256), # EC
            rgb(44/256,160/256,44/256),  # Pericyte
            rgb(23/256,190/256,207/256), # Blood
            rgb(197/256,168/256,206/256), # Neutrophils
            rgb(239/256,132/256,118/256), # GMPs
            rgb(249/256,219/256,149/256), # Macrophage
            rgb(140/256,205/256,191/256)  # T cell
            ) 

Pie <- as.data.frame(all_pie)
colnames(Pie) <- c('Cluster','Count')
Pie$Cluster <- factor(Pie$Cluster,
                          levels = Pie$Cluster)
Pie$Percent <- Pie$Count/sum(Pie$Count)

# pie(Pie$Percent, labels = with(Pie, paste0(round(Percent, 2) * 100, "%")), 
#     col = mycolor)

ggplot(Pie,aes(x = factor(1), y = Percent, fill = Cluster)) +
  geom_col(colour = "white")+
  coord_polar(theta = "y")+
  scale_fill_manual(values=color2)+
  geom_text(aes(label = paste0(round(Percent * 100, 2), "%")), 
            position = position_fill(vjust = 0.5)) +
  theme(
    panel.background = element_blank(),
    axis.title = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank()
  )
ggsave(paste("/hejian1/Scissors/AS_BR/Res_data/Figures/BR_All_Pie.pdf"),
       plot = last_plot(),
       device = cairo_pdf,width =5, height =5,dpi = 300)

```
